refactoring(operator): move mappers out from reconcillers
