+++ kroxylicious-operator/src/test/java/io/kroxylicious/kubernetes/operator/reconciler/kafkaproxy/KafkaProxyPrimaryToKafkaProtocolFilterSecondaryMapperTest.java
@@ -1,0 +1,93 @@
+/*
+ * Copyright Kroxylicious Authors.
+ *
+ * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
+ */
+
+package io.kroxylicious.kubernetes.operator.reconciler.kafkaproxy;
+
+import java.util.List;
+import java.util.Set;
+
+import org.junit.jupiter.api.Test;
+
+import io.fabric8.kubernetes.api.model.KubernetesResourceList;
+import io.fabric8.kubernetes.client.KubernetesClient;
+import io.fabric8.kubernetes.client.dsl.MixedOperation;
+import io.fabric8.kubernetes.client.dsl.Resource;
+import io.javaoperatorsdk.operator.api.reconciler.EventSourceContext;
+import io.javaoperatorsdk.operator.processing.event.ResourceID;
+import io.javaoperatorsdk.operator.processing.event.source.PrimaryToSecondaryMapper;
+
+import io.kroxylicious.kubernetes.api.common.FilterRef;
+import io.kroxylicious.kubernetes.api.v1alpha1.KafkaProxy;
+import io.kroxylicious.kubernetes.api.v1alpha1.KafkaProxyBuilder;
+import io.kroxylicious.kubernetes.api.v1alpha1.VirtualKafkaCluster;
+import io.kroxylicious.kubernetes.api.v1alpha1.VirtualKafkaClusterBuilder;
+
+import static io.kroxylicious.kubernetes.operator.ResourcesUtil.name;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
+
+class KafkaProxyPrimaryToKafkaProtocolFilterSecondaryMapperTest {
+
+    @Test
+    void proxyToFilterMapping() {
+        EventSourceContext<KafkaProxy> eventSourceContext = mock();
+        KubernetesClient client = mock();
+        when(eventSourceContext.getClient()).thenReturn(client);
+        MixedOperation<VirtualKafkaCluster, KubernetesResourceList<VirtualKafkaCluster>, Resource<VirtualKafkaCluster>> mockOperation = mock();
+        when(client.resources(VirtualKafkaCluster.class)).thenReturn(mockOperation);
+        KubernetesResourceList<VirtualKafkaCluster> mockList = mock();
+        when(mockOperation.list()).thenReturn(mockList);
+        when(mockOperation.inNamespace(any())).thenReturn(mockOperation);
+        String proxyName = "proxy";
+        String proxyNamespace = "test";
+        String filterName = "filter";
+        KafkaProxy proxy = new KafkaProxyBuilder().withNewMetadata().withName(proxyName).withNamespace(proxyNamespace).endMetadata().build();
+        VirtualKafkaCluster clusterForAnotherProxy = baseVirtualKafkaClusterBuilder(proxy, "cluster", 1L, 1L).editOrNewSpec()
+                .addNewFilterRef().withName(filterName).endFilterRef().endSpec().build();
+        when(mockList.getItems()).thenReturn(List.of(clusterForAnotherProxy));
+        PrimaryToSecondaryMapper<KafkaProxy> mapper = new KafkaProxyPrimaryToKafkaProtocolFilterSecondaryMapper(eventSourceContext);
+        Set<ResourceID> secondaryResourceIDs = mapper.toSecondaryResourceIDs(proxy);
+        assertThat(secondaryResourceIDs).containsExactly(new ResourceID(filterName, proxyNamespace));
+        verify(mockOperation).inNamespace(proxyNamespace);
+    }
+
+    @Test
+    void proxyToFilterMappingForClusterWithoutFilters() {
+        EventSourceContext<KafkaProxy> eventSourceContext = mock();
+        KubernetesClient client = mock();
+        when(eventSourceContext.getClient()).thenReturn(client);
+        KubernetesResourceList<VirtualKafkaCluster> mockList = MapperTestSupport.mockVirtualKafkaClusterListOperation(client);
+        KafkaProxy proxy = MapperTestSupport.buildProxy("proxy");
+        VirtualKafkaCluster clusterWithoutFilters = baseVirtualKafkaClusterBuilder(proxy, "cluster", 1L, 1L)
+                .editOrNewSpec()
+                .withFilterRefs((List<FilterRef>) null)
+                .endSpec()
+                .build();
+        when(mockList.getItems()).thenReturn(List.of(clusterWithoutFilters));
+
+        PrimaryToSecondaryMapper<KafkaProxy> mapper = new KafkaProxyPrimaryToKafkaProtocolFilterSecondaryMapper(eventSourceContext);
+        Set<ResourceID> secondaryResourceIDs = mapper.toSecondaryResourceIDs(proxy);
+        assertThat(secondaryResourceIDs).isEmpty();
+    }
+
+    private VirtualKafkaClusterBuilder baseVirtualKafkaClusterBuilder(KafkaProxy kafkaProxy, String name, long generation, long observedGeneration) {
+        return new VirtualKafkaClusterBuilder()
+                .withNewMetadata()
+                .withName(name)
+                .withGeneration(generation)
+                .endMetadata()
+                .withNewSpec()
+                .withNewProxyRef()
+                .withName(name(kafkaProxy))
+                .endProxyRef().endSpec()
+                .editStatus().withObservedGeneration(observedGeneration)
+                .endStatus();
+    }
+
+}
