+++ kroxylicious-operator/src/test/java/io/kroxylicious/kubernetes/operator/reconciler/kafkaproxy/KafkaProxyToVirtualKafkaClusterPrimaryToSecondaryMapperTest.java
@@ -1,0 +1,70 @@
+/*
+ * Copyright Kroxylicious Authors.
+ *
+ * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
+ */
+
+package io.kroxylicious.kubernetes.operator.reconciler.kafkaproxy;
+
+import java.util.List;
+import java.util.Set;
+
+import org.junit.jupiter.api.Test;
+
+import io.fabric8.kubernetes.api.model.KubernetesResourceList;
+import io.fabric8.kubernetes.client.KubernetesClient;
+import io.fabric8.kubernetes.client.dsl.MixedOperation;
+import io.fabric8.kubernetes.client.dsl.Resource;
+import io.javaoperatorsdk.operator.api.reconciler.EventSourceContext;
+import io.javaoperatorsdk.operator.processing.event.ResourceID;
+import io.javaoperatorsdk.operator.processing.event.source.PrimaryToSecondaryMapper;
+
+import io.kroxylicious.kubernetes.api.v1alpha1.KafkaProxy;
+import io.kroxylicious.kubernetes.api.v1alpha1.VirtualKafkaCluster;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
+
+class KafkaProxyToVirtualKafkaClusterPrimaryToSecondaryMapperTest {
+    @Test
+    void proxyToClusterMapper() {
+        EventSourceContext<KafkaProxy> eventSourceContext = mock();
+        KubernetesClient client = mock();
+        when(eventSourceContext.getClient()).thenReturn(client);
+        MixedOperation<VirtualKafkaCluster, KubernetesResourceList<VirtualKafkaCluster>, Resource<VirtualKafkaCluster>> mockOperation = mock();
+        when(client.resources(VirtualKafkaCluster.class)).thenReturn(mockOperation);
+        KubernetesResourceList<VirtualKafkaCluster> mockList = mock();
+        when(mockOperation.list()).thenReturn(mockList);
+        when(mockOperation.inNamespace(any())).thenReturn(mockOperation);
+        KafkaProxy proxy = MapperTestSupport.buildProxy("proxy");
+        VirtualKafkaCluster cluster = MapperTestSupport.baseVirtualKafkaClusterBuilder(proxy, "cluster", 1L, 1L).build();
+        VirtualKafkaCluster clusterForAnotherProxy = MapperTestSupport
+                .baseVirtualKafkaClusterBuilder(MapperTestSupport.buildProxy("anotherproxy"), "anothercluster", 1L, 1L).build();
+        when(mockList.getItems()).thenReturn(List.of(cluster, clusterForAnotherProxy));
+        PrimaryToSecondaryMapper<KafkaProxy> mapper = new KafkaProxyPrimaryToVirtualKafkaClusterSecondaryMapper(eventSourceContext);
+        Set<ResourceID> secondaryResourceIDs = mapper.toSecondaryResourceIDs(proxy);
+        assertThat(secondaryResourceIDs).containsExactly(ResourceID.fromResource(cluster));
+    }
+
+    @Test
+    void proxyToClusterMapper_NoMatchedClusters() {
+        EventSourceContext<KafkaProxy> eventSourceContext = mock();
+        KubernetesClient client = mock();
+        when(eventSourceContext.getClient()).thenReturn(client);
+        MixedOperation<VirtualKafkaCluster, KubernetesResourceList<VirtualKafkaCluster>, Resource<VirtualKafkaCluster>> mockOperation = mock();
+        when(client.resources(VirtualKafkaCluster.class)).thenReturn(mockOperation);
+        KubernetesResourceList<VirtualKafkaCluster> mockList = mock();
+        when(mockOperation.list()).thenReturn(mockList);
+        when(mockOperation.inNamespace(any())).thenReturn(mockOperation);
+        VirtualKafkaCluster clusterForAnotherProxy = MapperTestSupport.baseVirtualKafkaClusterBuilder(MapperTestSupport.buildProxy("anotherProxy"), "cluster", 1L, 1L)
+                .build();
+        when(mockList.getItems()).thenReturn(List.of(clusterForAnotherProxy));
+        PrimaryToSecondaryMapper<KafkaProxy> mapper = new KafkaProxyPrimaryToVirtualKafkaClusterSecondaryMapper(eventSourceContext);
+        KafkaProxy proxy = MapperTestSupport.buildProxy("proxy");
+        Set<ResourceID> secondaryResourceIDs = mapper.toSecondaryResourceIDs(proxy);
+        assertThat(secondaryResourceIDs).isEmpty();
+    }
+
+}
